Refactor validation and save logic in one method (#1096)
Summary:
# Before submitting

- [ ] Was this discussed/approved via a Github issue? (no need for typos, doc improvements)
- [ ] Did you read the [contributor guideline](https://github.com/pytorch/fairseq/blob/master/CONTRIBUTING.md)?
- [ ] Did you make sure to update the docs?
- [ ] Did you write any new necessary tests?

## What does this PR do?
Related to issue: https://github.com/fairinternal/fairseq-py/issues/1089

I found it hard to read the save and evaluation logic before (and it was probably hard to maintain as well).
This PR would also fixes some inconsistencies.
The same behaviour is applied for validation during epoch as for validation as the end of epoch. This is useful for patience for instance, the early stopping happens in terms of number of validations (and can therefore happen during one epoch) instead of in number of full epochs, which can be problematic when epochs are very long.
Pull Request resolved: https://github.com/fairinternal/fairseq-py/pull/1096

Reviewed By: myleott

Differential Revision: D20557186

Pulled By: louismartin

fbshipit-source-id: 54129c357a5c5cd4b5c26b7dd29f23cc2c6c40d4