Add --load-checkpoint-on-all-dp-ranks (#1478)
Summary:
A recent commit [1] made it so that checkpoints are loaded on rank 0 and then broadcast to other workers. This is valuable to reduce I/O and especially helpful when training with optimizer state sharding, so we don't need to load redundant optimizer state on every worker.

This diff adds a new option (`--load-checkpoint-on-all-dp-ranks`) that optionally reverts to the old behavior.

[1] https://github.com/pytorch/fairseq/commit/45854b758d9f4a13a6485e2e3065e5dd00b4e681

Pull Request resolved: https://github.com/fairinternal/fairseq-py/pull/1478

Test Plan: Imported from OSS

Reviewed By: theweiho

Differential Revision: D25291595

Pulled By: myleott

fbshipit-source-id: 57b521c5e6a48f08140f9527162072ea1d4066db