delegate namespace conversion to DC
Summary: `populate_dataclass` is very basic in how it populates the dataclass. We might want more specific behaviour for some config dataclasses (like the hierarchical behaviour in TransformerConfig, see rest of stack). This diff move the populate logic to a `from_namespace` method in `FairseqDataclass` so that the a specific Dataclass can reimplement it.

Reviewed By: myleott

Differential Revision: D29521388

fbshipit-source-id: f3a6dc80e4ddfc9563c6e85c37c563173f193f4d